user_service/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
Django==4.2.21
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
pyparsing==3.2.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
django-cors-headers==4.7.0
google-auth==2.40.2